# sigpcb-base-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Base",
        "description": "Gestion des API de base",
        "contact": {
            "name": "Support client",
            "email": "contact@exemple.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/anatt-base/salle-compos": {
            "get": {
                "tags": [
                    "SalleCompos"
                ],
                "summary": "Récupérer la liste des salles de composition",
                "description": "Récupère une liste de tous les salles de composition enregistrées dans la base de données",
                "operationId": "getAllSalleCompos",
                "responses": {
                    "200": {
                        "description": "La liste des salles de composition récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID de la salle de composition",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Nom de la salle de composition",
                                                "type": "string"
                                            },
                                            "annexe_anatt_id": {
                                                "description": "ID de l'annexe anatt auquel appartient la nouvelle salle créée",
                                                "type": "integer"
                                            },
                                            "contenance": {
                                                "description": "Contenance de la salle de composition",
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SalleCompos"
                ],
                "summary": "Crée une nouvelle salle de composition",
                "description": "Crée une nouvelle salle de composition enregistrée dans la base de données",
                "operationId": "createSalleCompo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom de la salle de composition",
                                        "type": "string"
                                    },
                                    "annexe_anatt_id": {
                                        "description": "ID de l'annexe anatt auquel appartient la nouvelle salle créée",
                                        "type": "integer"
                                    },
                                    "contenance": {
                                        "description": "Contenance de la salle de composition",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nouvelle salle de composition créée"
                    }
                }
            }
        },
        "/api/anatt-base/salle-compo/multiple": {
            "post": {
                "tags": [
                    "SalleCompos"
                ],
                "summary": "Crée plusieurs salles de composition",
                "description": "Crée plusieurs salles de composition enregistrées dans la base de données pour une annexe ANATT donnée",
                "operationId": "createMultipleSalleCompo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "annexe_anatt_id": {
                                        "description": "ID de l'annexe ANATT à laquelle appartiennent les nouvelles salles créées",
                                        "type": "integer"
                                    },
                                    "salle_comp_groupes": {
                                        "description": "Tableau contenant les informations sur les salles de composition à créer",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "description": "Nom de la salle de composition",
                                                    "type": "string"
                                                },
                                                "contenance": {
                                                    "description": "Contenance de la salle de composition",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nouvelles salles de composition créées"
                    },
                    "422": {
                        "description": "Erreur de validation des données d'entrée"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            }
        },
        "/api/anatt-base/salle-compos/{id}": {
            "get": {
                "tags": [
                    "SalleCompos"
                ],
                "summary": "Récupère une salle de composition par ID",
                "description": "Récupère une salle de composition enregistré dans la base de données en spécifiant son ID",
                "operationId": "getSalleCompoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle de composition à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salle de composition récupérée avec succès"
                    },
                    "404": {
                        "description": "Salle de composition non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "SalleCompos"
                ],
                "summary": "Met à jour une salle de composition existante",
                "description": "Met à jour une salle de composition existante dans la base de données",
                "operationId": "updateSalleCompo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle de composition à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom de la salle de composition",
                                        "type": "string"
                                    },
                                    "annexe_anatt_id": {
                                        "description": "ID de l'annexe antt auquel appartient la nouvelle salle créée",
                                        "type": "integer"
                                    },
                                    "contenance": {
                                        "description": "Contenance de la salle de composition",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salle de composition mise à jour avec succès"
                    },
                    "404": {
                        "description": "Salle de composition non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SalleCompos"
                ],
                "summary": "Supprime une salle de composition",
                "description": "Supprime une salle de composition de la base de données",
                "operationId": "deleteSalleCompo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle de composition à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salle de composition supprimée avec succès"
                    },
                    "404": {
                        "description": "Salle de composition non trouvée"
                    }
                }
            }
        },
        "/api/anatt-base/annexeanatt-salle-compos/{annexe_anatt_id}": {
            "get": {
                "tags": [
                    "SalleCompos"
                ],
                "summary": "Récupère les salles d'une annexe anatt",
                "description": "Récupère les salles d'une annexe anatt enregistré dans la base de données en spécifiant son ID",
                "operationId": "getAnnexeAnattSalleCompoById",
                "parameters": [
                    {
                        "name": "annexe_anatt_id",
                        "in": "path",
                        "description": "ID de l'annexe anatt dont on veut récuperer la salle",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salle de composition récupérée avec succès"
                    },
                    "404": {
                        "description": "Salle de composition non trouvée"
                    }
                }
            }
        },
...
    }
}